Models rearranged and Context file updated
