+++ InternshipBE/DAL/Entities/Feedback.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DAL.Models
+namespace DAL.Entities
 {
     public class Feedback
     {
