+++ InternshipBE/DAL/Entities/Role.cs
@@ -4,15 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DAL.Models
+namespace DAL.Entities
 {
-    public class InterviewInvite
+    public class Role
     {
         public int Id { get; set; }
 
-        public int UserCandidateId { get; set; }
-
-        public DateTime ContactDate { get; set; }
+        public string Name { get; set; }
 
     }
 }
