+++ InternshipBE/DAL/Entities/UserCandidate.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DAL.Models
+namespace DAL.Entities
 {
     public class UserCandidate
     {
