--- InternshipBE/DAL/SandboxContext.cs
@@ -1,36 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using DAL.Models;
-using Microsoft.EntityFrameworkCore;
-
-namespace DAL
-{
-    
-    public class SandboxContext: DbContext
-    {
-        public SandboxContext(DbContextOptions<SandboxContext> options): base(options)
-        { 
-               
-        }
-        public DbSet<User> Users { get; set; }
-
-        public DbSet<Feedback> Feedbacks { get; set; }
-
-        public DbSet<Evaluation> Evaluations { get; set; }
-
-        public DbSet<Skill> Skills { get; set; }
-
-        public DbSet<Role> Roles { get; set; }
-
-        public DbSet<InterviewInvite> InterviewInvites { get; set; }
-
-        public DbSet<Candidate> Candidates { get; set; }
-        public DbSet<Internship> Internships {  get; set; }
-        public DbSet<InternshipStack> InternshipStacks { get; set; }
-        public DbSet<UserCandidate> UserCandidates { get; set; }
-
-    }
-}
