+++ InternshipBE/WebApi/appsettings.json
@@ -7,7 +7,7 @@
     "Key": "732a81a384ca610d4bd9d13baf3feb10",
     "Issuer": "ExadelAuthServer",
     "Audience": "ExadelTeam",
-    "LifeTimeInDays": 30 
+    "LifeTimeInDays": 30
   },
 
   "Logging": {
@@ -18,7 +18,6 @@
     }
   },
   "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=SandboxDB;Trusted_Connection=True;MultipleActiveResultSets=true"
-  }
+
+ 
 }
